Move Haskell-types instances of SpecTranslate from Conway.Base to Base
