+++ libs/cardano-ledger-conformance/src/Test/Cardano/Ledger/Conformance/SpecTranslate/Base.hs
@@ -27,15 +27,30 @@ module Test.Cardano.Ledger.Conformance.SpecTranslate.Base (
   toSpecRep_,
 ) where
 
-import Cardano.Ledger.BaseTypes (Inject (..))
+import Cardano.Ledger.Binary (Sized (..))
+import Cardano.Ledger.BaseTypes (Inject (..), NonNegativeInterval, UnitInterval, unboundRational)
+import Cardano.Ledger.Compactible (Compactible (..), fromCompact)
+import Data.List.NonEmpty (NonEmpty)
 import Control.DeepSeq (NFData)
 import Control.Monad.Except (ExceptT, MonadError (..), runExceptT)
 import Control.Monad.Reader (MonadReader (..), Reader, asks, runReader)
+import Data.Bitraversable (bimapM)
+import Data.Foldable (Foldable (..))
 import Data.Kind (Type)
-import Data.List.NonEmpty (NonEmpty)
+import Data.Maybe.Strict (StrictMaybe (..), strictMaybeToMaybe)
+import Data.Map (Map)
+import qualified Data.Map as Map
+import Data.OMap.Strict (OMap)
+import qualified Data.OMap.Strict as OMap
+import Data.OSet.Strict (OSet)
+import Data.Sequence (Seq)
+import Data.Sequence.Strict (StrictSeq)
+import Data.Set (Set)
+import qualified Data.Set as Set
 import Data.Text (Text)
 import qualified Data.Text as T
-import Data.Void (Void)
+import Data.Void (Void, absurd)
+import Data.Word (Word16, Word32, Word64)
 import GHC.Generics (Generic (..), K1 (..), M1 (..), U1 (..), V1, (:*:) (..), (:+:) (..))
 import qualified MAlonzo.Code.Ledger.Foreign.API as Agda
 import Test.Cardano.Ledger.TreeDiff (Expr (..), ToExpr (..))
@@ -51,6 +66,127 @@ class SpecTranslate ctx a where
   type SpecRep a :: Type
 
   toSpecRep :: a -> SpecTransM ctx (SpecRep a)
+
+instance SpecTranslate ctx () where
+  type SpecRep () = ()
+
+  toSpecRep = pure
+
+instance SpecTranslate ctx Bool where
+  type SpecRep Bool = Bool
+
+  toSpecRep = pure
+
+instance SpecTranslate ctx Integer where
+  type SpecRep Integer = Integer
+
+  toSpecRep = pure
+
+instance SpecTranslate ctx Void where
+  type SpecRep Void = Void
+
+  toSpecRep = absurd
+
+instance SpecTranslate ctx Word16 where
+  type SpecRep Word16 = Integer
+
+  toSpecRep = pure . toInteger
+
+instance SpecTranslate ctx Word32 where
+  type SpecRep Word32 = Integer
+
+  toSpecRep = pure . toInteger
+
+instance SpecTranslate ctx Word64 where
+  type SpecRep Word64 = Integer
+
+  toSpecRep = pure . toInteger
+
+instance
+  ( SpecTranslate ctx a
+  , SpecTranslate ctx b
+  ) =>
+  SpecTranslate ctx (a, b)
+  where
+  type SpecRep (a, b) = (SpecRep a, SpecRep b)
+
+  toSpecRep (x, y) = (,) <$> toSpecRep x <*> toSpecRep y
+
+instance SpecTranslate ctx a => SpecTranslate ctx [a] where
+  type SpecRep [a] = [SpecRep a]
+
+  toSpecRep = traverse toSpecRep
+
+instance SpecTranslate ctx a => SpecTranslate ctx (StrictMaybe a) where
+  type SpecRep (StrictMaybe a) = Maybe (SpecRep a)
+
+  toSpecRep = toSpecRep . strictMaybeToMaybe
+
+instance SpecTranslate ctx a => SpecTranslate ctx (Maybe a) where
+  type SpecRep (Maybe a) = Maybe (SpecRep a)
+
+  toSpecRep = traverse toSpecRep
+
+instance SpecTranslate ctx a => SpecTranslate ctx (StrictSeq a) where
+  type SpecRep (StrictSeq a) = [SpecRep a]
+
+  toSpecRep = traverse toSpecRep . toList
+
+instance SpecTranslate ctx a => SpecTranslate ctx (Seq a) where
+  type SpecRep (Seq a) = [SpecRep a]
+
+  toSpecRep = traverse toSpecRep . toList
+
+instance SpecTranslate ctx a => SpecTranslate ctx (OSet a) where
+  type SpecRep (OSet a) = [SpecRep a]
+
+  toSpecRep = traverse toSpecRep . toList
+
+instance
+  ( SpecTranslate ctx k
+  , SpecTranslate ctx v
+  , Ord k
+  ) =>
+  SpecTranslate ctx (OMap k v)
+  where
+  type SpecRep (OMap k v) = [(SpecRep k, SpecRep v)]
+
+  toSpecRep = traverse (bimapM toSpecRep toSpecRep) . OMap.assocList
+
+instance (SpecTranslate ctx a, Compactible a) => SpecTranslate ctx (CompactForm a) where
+  type SpecRep (CompactForm a) = SpecRep a
+
+  toSpecRep = toSpecRep . fromCompact
+
+instance SpecTranslate ctx a => SpecTranslate ctx (Sized a) where
+  type SpecRep (Sized a) = SpecRep a
+
+  toSpecRep (Sized x _) = toSpecRep x
+
+instance SpecTranslate ctx a => SpecTranslate ctx (Set a) where
+  type SpecRep (Set a) = Agda.HSSet (SpecRep a)
+
+  toSpecRep = fmap Agda.MkHSSet . traverse toSpecRep . Set.toList
+
+instance SpecTranslate ctx UnitInterval where
+  type SpecRep UnitInterval = Agda.Rational
+
+  toSpecRep = pure . unboundRational
+
+instance SpecTranslate ctx NonNegativeInterval where
+  type SpecRep NonNegativeInterval = Agda.Rational
+
+  toSpecRep = pure . unboundRational
+
+instance
+  ( SpecTranslate ctx k
+  , SpecTranslate ctx v
+  ) =>
+  SpecTranslate ctx (Map k v)
+  where
+  type SpecRep (Map k v) = Agda.HSMap (SpecRep k) (SpecRep v)
+
+  toSpecRep = fmap Agda.MkHSMap . traverse (bimapM toSpecRep toSpecRep) . Map.toList
 
 class GSpecNormalize f where
   genericSpecNormalize :: f a -> f a
