+++ libs/cardano-ledger-conformance/src/Test/Cardano/Ledger/Conformance/SpecTranslate/Conway/Base.hs
@@ -47,9 +47,7 @@ import Cardano.Ledger.Alonzo.Scripts (AlonzoPlutusPurpose (..))
 import Cardano.Ledger.Alonzo.TxWits (AlonzoTxWits (..), Redeemers (..), TxDats (..), unTxDats)
 import Cardano.Ledger.Babbage.TxOut (BabbageTxOut (..))
 import Cardano.Ledger.BaseTypes
-import Cardano.Ledger.Binary (Sized (..))
-import Cardano.Ledger.Coin (Coin (..), CompactForm)
-import Cardano.Ledger.Compactible (Compactible, fromCompact)
+import Cardano.Ledger.Coin (Coin (..))
 import Cardano.Ledger.Conway.Core
 import Cardano.Ledger.Conway.Governance
 import Cardano.Ledger.Conway.PParams (ConwayEraPParams (..), ConwayPParams (..), THKD (..))
@@ -81,16 +79,9 @@ import Data.List (sortOn)
 import Data.Map.Strict (Map)
 import qualified Data.Map.Strict as Map
 import Data.OMap.Strict (OMap)
-import qualified Data.OMap.Strict as OMap
-import Data.OSet.Strict (OSet)
-import Data.Sequence (Seq)
-import Data.Sequence.Strict (StrictSeq)
-import Data.Set (Set)
 import qualified Data.Set as Set
 import qualified Data.Text as T
 import Data.Traversable (forM)
-import Data.Void (Void, absurd)
-import Data.Word (Word16, Word32, Word64)
 import qualified GHC.Exts as Exts
 import Lens.Micro
 import Lens.Micro.Extras (view)
@@ -102,20 +93,6 @@ import Test.Cardano.Ledger.Conformance (
   hashToInteger,
  )
 import Test.Cardano.Ledger.Conway.TreeDiff (ToExpr (..), showExpr)
-
-instance SpecTranslate ctx Void where
-  type SpecRep Void = Void
-
-  toSpecRep = absurd
-
-instance SpecTranslate ctx () where
-  type SpecRep () = ()
-  toSpecRep = pure
-
-instance SpecTranslate ctx a => SpecTranslate ctx [a] where
-  type SpecRep [a] = [SpecRep a]
-
-  toSpecRep = traverse toSpecRep
 
 instance SpecTranslate ctx TxIx where
   type SpecRep TxIx = Integer
@@ -249,11 +226,6 @@ instance
     datum' <- toSpecRep datum
     script' <- toSpecRep script
     pure (addr', (val', (datum', script')))
-
-instance SpecTranslate ctx Integer where
-  type SpecRep Integer = Integer
-
-  toSpecRep = pure
 
 deriving instance SpecTranslate ctx Coin
 
@@ -395,26 +367,6 @@ instance
 
   toSpecRep (PParams x) = toSpecRep x
 
-instance SpecTranslate ctx a => SpecTranslate ctx (Set a) where
-  type SpecRep (Set a) = Agda.HSSet (SpecRep a)
-
-  toSpecRep = fmap Agda.MkHSSet . traverse toSpecRep . Set.toList
-
-instance SpecTranslate ctx a => SpecTranslate ctx (StrictSeq a) where
-  type SpecRep (StrictSeq a) = [SpecRep a]
-
-  toSpecRep = traverse toSpecRep . toList
-
-instance SpecTranslate ctx a => SpecTranslate ctx (Seq a) where
-  type SpecRep (Seq a) = [SpecRep a]
-
-  toSpecRep = traverse toSpecRep . toList
-
-instance SpecTranslate ctx a => SpecTranslate ctx (Sized a) where
-  type SpecRep (Sized a) = SpecRep a
-
-  toSpecRep (Sized x _) = toSpecRep x
-
 instance SpecTranslate ctx ValidityInterval where
   type SpecRep ValidityInterval = (Maybe Integer, Maybe Integer)
 
@@ -460,31 +412,6 @@ instance Era era => SpecTranslate ctx (TxDats era) where
 
   toSpecRep = fmap Agda.MkHSSet . traverse (toSpecRep . snd) . Map.toList . unTxDats
 
-instance
-  ( SpecTranslate ctx k
-  , SpecTranslate ctx v
-  , Ord k
-  ) =>
-  SpecTranslate ctx (Map k v)
-  where
-  type SpecRep (Map k v) = Agda.HSMap (SpecRep k) (SpecRep v)
-
-  toSpecRep =
-    fmap Agda.MkHSMap
-      . traverse (bimapM toSpecRep toSpecRep)
-      . sortOn fst
-      . Map.toList
-
-instance SpecTranslate ctx Word64 where
-  type SpecRep Word64 = Integer
-
-  toSpecRep = pure . toInteger
-
-instance SpecTranslate ctx Word32 where
-  type SpecRep Word32 = Integer
-
-  toSpecRep = pure . toInteger
-
 instance SpecTranslate ctx (AlonzoPlutusPurpose AsIx era) where
   type SpecRep (AlonzoPlutusPurpose AsIx era) = Agda.RdmrPtr
 
@@ -504,16 +431,6 @@ instance SpecTranslate ctx (ConwayPlutusPurpose AsIx era) where
     ConwayRewarding (AsIx i) -> pure (Agda.Rewrd, toInteger i)
     ConwayVoting (AsIx i) -> pure (Agda.Vote, toInteger i)
     ConwayProposing (AsIx i) -> pure (Agda.Propose, toInteger i)
-
-instance
-  ( SpecTranslate ctx a
-  , SpecTranslate ctx b
-  ) =>
-  SpecTranslate ctx (a, b)
-  where
-  type SpecRep (a, b) = (SpecRep a, SpecRep b)
-
-  toSpecRep (x, y) = (,) <$> toSpecRep x <*> toSpecRep y
 
 instance Era era => SpecTranslate ctx (Data era) where
   type SpecRep (Data era) = Agda.DataHash
@@ -551,16 +468,6 @@ instance
       <*> toSpecRep (txrdmrs x)
     where
       txWitsMap = toList (txwitsVKey x)
-
-instance SpecTranslate ctx a => SpecTranslate ctx (StrictMaybe a) where
-  type SpecRep (StrictMaybe a) = Maybe (SpecRep a)
-
-  toSpecRep = toSpecRep . strictMaybeToMaybe
-
-instance SpecTranslate ctx a => SpecTranslate ctx (Maybe a) where
-  type SpecRep (Maybe a) = Maybe (SpecRep a)
-
-  toSpecRep = traverse toSpecRep
 
 instance Era era => SpecTranslate ctx (AlonzoTxAuxData era) where
   type SpecRep (AlonzoTxAuxData era) = Agda.AuxiliaryData
@@ -639,11 +546,6 @@ instance SpecTranslate ctx (GovPurposeId r) where
 
   toSpecRep (GovPurposeId gaId) = toSpecRep gaId
 
-instance SpecTranslate ctx UnitInterval where
-  type SpecRep UnitInterval = Agda.Rational
-
-  toSpecRep = pure . unboundRational
-
 instance SpecTranslate ctx (Committee era) where
   type SpecRep (Committee era) = (Agda.HSMap Agda.Credential Agda.Epoch, Agda.Rational)
 
@@ -718,16 +620,6 @@ instance SpecTranslate ctx (VotingProcedures era) where
                   <*> toSpecRep (vProcAnchor votingProcedure)
               )
           <*> m
-
-instance SpecTranslate ctx Word16 where
-  type SpecRep Word16 = Integer
-
-  toSpecRep = pure . toInteger
-
-instance SpecTranslate ctx NonNegativeInterval where
-  type SpecRep NonNegativeInterval = Agda.Rational
-
-  toSpecRep = pure . unboundRational
 
 instance SpecTranslate ctx (ConwayPParams StrictMaybe era) where
   type SpecRep (ConwayPParams StrictMaybe era) = Agda.PParamsUpdate
@@ -810,22 +702,6 @@ instance
       <$> toSpecRep h
       <*> toSpecRep policy
   toSpecRep InfoAction = pure Agda.Info
-
-instance SpecTranslate ctx a => SpecTranslate ctx (OSet a) where
-  type SpecRep (OSet a) = [SpecRep a]
-
-  toSpecRep = traverse toSpecRep . toList
-
-instance
-  ( SpecTranslate ctx k
-  , SpecTranslate ctx v
-  , Ord k
-  ) =>
-  SpecTranslate ctx (OMap k v)
-  where
-  type SpecRep (OMap k v) = [(SpecRep k, SpecRep v)]
-
-  toSpecRep = traverse (bimapM toSpecRep toSpecRep) . OMap.assocList
 
 instance
   ( EraPParams era
@@ -935,10 +811,7 @@ instance SpecTranslate ctx MaryValue where
 
   toSpecRep = toSpecRep . coin
 
-instance (SpecTranslate ctx a, Compactible a) => SpecTranslate ctx (CompactForm a) where
-  type SpecRep (CompactForm a) = SpecRep a
 
-  toSpecRep = toSpecRep . fromCompact
 
 instance SpecTranslate ctx CommitteeAuthorization where
   type
@@ -991,11 +864,6 @@ instance
       <*> toSpecRep treasury
       <*> toSpecRep (Map.mapWithKey stakePoolStateToPoolParams reStakePools)
       <*> toSpecRep (Map.mapMaybe (^. dRepDelegationAccountStateL) (reAccounts ^. accountsMapL))
-
-instance SpecTranslate ctx Bool where
-  type SpecRep Bool = Bool
-
-  toSpecRep = pure
 
 instance
   ( EraPParams era
